# swagger.yaml
openapi: 3.0.0
info:
  title: API do Meu Comercio
  version: 1.0.0
  description: Documentação da API do Meu Comercio
security: []
paths:
  /api/admin:
    get:
      summary: Verifica se já existe um administrador
      tags:
        - Admin
      description: Retorna se existe algum usuário com a role "Administrador".
      responses:
        '200':
          description: Resultado da verificação
          content:
            application/json:
              schema:
                type: object
                properties:
                  hasAdmin:
                    type: boolean
                    example: true
        '500':
          description: Erro interno ao listar usuários
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Failed to list users
    post:
      summary: Cria o primeiro administrador
      tags:
        - Admin
      description: Cria um usuário com role "Administrador" apenas se não existir outro.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  example: admin@empresa.com
                password:
                  type: string
                  example: SenhaSegura123
                displayName:
                  type: string
                  example: Administrador
      responses:
        '200':
          description: Administrador criado com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  uid:
                    type: string
                    example: abc123xyz
                  email:
                    type: string
                    example: admin@empresa.com
                  message:
                    type: string
                    example: Administrador criado com sucesso!
        '400':
          description: Erro de validação (email/senha ausente ou admin já existente)
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Administrador já existe
        '500':
          description: Erro interno ao criar o usuário
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Erro ao criar usuário no Firebase
  /api/vendas/{id}:
    get:
      summary: Recupera uma venda pelo ID
      tags:
        - Vendas
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID da venda a ser consultada
      responses:
        '200':
          description: Venda encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: v1
                  usuarioId:
                    type: string
                    example: u1
                  itens:
                    type: array
                    items:
                      type: object
                      properties:
                        produtoId:
                          type: string
                          example: p1
                        quantidade:
                          type: integer
                          example: 2
                        precoUnitario:
                          type: number
                          example: 15.5
                  valorTotal:
                    type: number
                    example: 31
                  criadoEm:
                    type: string
                    format: date-time
                    example: '2025-11-12T10:00:00.000Z'
        '404':
          description: Venda não encontrada
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Venda não encontrada
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Erro interno
    delete:
      summary: Remove uma venda pelo ID
      tags:
        - Vendas
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: ID da venda a ser removida
      responses:
        '200':
          description: Venda removida com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Erro interno
  /api/vendas:
    get:
      summary: Lista todas as vendas
      tags:
        - Vendas
      description: Recupera a lista de vendas registradas.
      responses:
        '200':
          description: Lista de vendas
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                      example: v1
                    usuarioId:
                      type: string
                      example: u1
                    formaPagamento:
                      type: string
                      example: Dinheiro
                    itens:
                      type: array
                      items:
                        type: object
                        properties:
                          produtoId:
                            type: string
                            example: p1
                          quantidade:
                            type: integer
                            example: 2
                          precoUnitario:
                            type: number
                            example: 15.5
                    valorTotal:
                      type: number
                      example: 31
                    criadoEm:
                      type: string
                      format: date-time
                      example: '2025-11-12T10:00:00.000Z'
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Erro interno
    post:
      summary: Cria uma nova venda
      tags:
        - Vendas
      description: Salva uma venda com itens e registra movimentos de estoque.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - usuarioId
                - itens
                - valorTotal
              properties:
                usuarioId:
                  type: string
                  example: u1
                formaPagamento:
                  type: string
                  enum:
                    - Dinheiro
                    - Cartão
                    - Pix
                  example: Pix
                itens:
                  type: array
                  items:
                    type: object
                    properties:
                      produtoId:
                        type: string
                        example: p1
                      quantidade:
                        type: integer
                        example: 2
                      precoUnitario:
                        type: number
                        example: 15.5
                valorTotal:
                  type: number
                  example: 31
      responses:
        '201':
          description: Venda criada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    example: v1
                  usuarioId:
                    type: string
                    example: u1
                  message:
                    type: string
                    example: Venda criada
        '400':
          description: Erro de validação
          content:
            application/json:
              schema:
                type: object
                properties:
                  errors:
                    type: object
        '500':
          description: Erro interno
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Erro interno
components: {}
tags:
  - name: Admin
    description: Gerenciamento do usuário administrador
  - name: Vendas
    description: Operações relacionadas a vendas